Add-remove nuimo feature (#252)
* Update frontend, backend and nuimo_app for Add-Remove multiple nuimos

* Fix scenario when active_component is removed

* Add test cases for complete coverage and adapt the existent ones to the latest changes

* Restart nuimo_app and bluetooth daemon every time a nuimo is added due to issues with multi-processes, adapted test cases accordingly

* Change trace level to info and enable restart in case of removed nuimo

* Retry connection in case of connection-failed during nuimo discovery

* Add swipe-to-delete feature

* Add NuimosMenu.js react file

* Remove debug trace

* Restore trace level

* Remove commented code

* Update review related code

* Add count for added-nuimo

* Correct comment logs

+++ senic_hub/backend/commands.py
@@ -15,6 +15,8 @@ COMPONENT_FOR_TYPE = {
 
 logger = logging.getLogger(__name__)
 
+import os.path
+
 
 def create_configuration_files_and_restart_apps(settings):
     # generate homeassistant config & restart supervisor app
@@ -34,9 +36,25 @@ def create_configuration_files_and_restart_apps(settings):
         nuimo_mac_address = f.readline().strip()
 
     nuimo_app_config_file_path = settings['nuimo_app_config_path']
-    with open(nuimo_app_config_file_path, 'w') as f:
-        config = generate_nuimo_app_configuration(nuimo_mac_address, devices)
-        yaml.dump(config, f, default_flow_style=False)
+
+    if os.path.isfile(nuimo_app_config_file_path) is False:
+        with open(nuimo_app_config_file_path, 'w') as f:
+            logger.info("file is empty")
+            config = generate_nuimo_app_configuration(nuimo_mac_address, devices)
+            yaml.dump(config, f, default_flow_style=False)
+    else:
+        with open(nuimo_app_config_file_path, 'r+') as f:
+            config = yaml.load(f) or None
+            logger.info("file has another configured nuimo")
+
+            components = [create_component(d) for d in devices if d["authenticated"]]
+            config['nuimos'][nuimo_mac_address] = {
+                'name': 'My Nuimo ' + str(len(config['nuimos']) + 1),
+                'components': components,
+            }
+            f.seek(0)  # We want to overwrite the config file with the new configuration
+            f.truncate()
+            yaml.dump(config, f, default_flow_style=False)
 
     supervisor.restart_program('nuimo_app')
